requirements.txt
dvc==2.8.3
flake8==4.0.1
matplotlib==3.5.0
mypy==0.910
pandas==1.3.4
pylint==2.11.1
pyyaml==6.0